+++ src/data/components/Menu.js
@@ -1,0 +1,17 @@
+export default {
+  name: 'Menu',
+  supports: {
+    bootstrap3: 'http://getbootstrap.com/components/#list-group-linked',
+    bootstrap4: 'https://v4-alpha.getbootstrap.com/components/list-group/#links-and-buttons',
+    bulma: 'http://bulma.io/documentation/components/menu/',
+    foundation5: 'http://foundation.zurb.com/sites/docs/v/5.5.3/components/sidenav.html',
+    foundation6: 'http://foundation.zurb.com/sites/docs/menu.html',
+    ink: 'http://ink.sapo.pt/ui-elements/navigation/#menus',
+    uikit2: 'https://getuikit.com/v2/docs/nav.html',
+    uikit3: 'https://getuikit.com/docs/nav',
+    pure: 'https://purecss.io/menus/',
+    primer: 'http://primercss.io/nav/',
+    semantic: 'https://semantic-ui.com/collections/menu.html',
+    spectre: 'https://picturepan2.github.io/spectre/components.html#menus'
+  }
+};