+++ src/data/components/Navbar.js
@@ -1,15 +1,15 @@
 export default {
-  name: 'Navbar / Navigation',
+  name: 'Navbar',
   supports: {
     bootstrap3: 'http://getbootstrap.com/components/#navbar',
     bootstrap4: 'https://v4-alpha.getbootstrap.com/components/navbar/',
     bulma: 'http://bulma.io/documentation/components/navbar/',
+    foundation5: 'http://foundation.zurb.com/sites/docs/v/5.5.3/components/topbar.html',
+    foundation6: 'http://foundation.zurb.com/sites/docs/top-bar.html',
     ink: 'http://ink.sapo.pt/ui-elements/navigation/',
     materialize: 'http://materializecss.com/navbar.html',
     uikit2: 'https://getuikit.com/v2/docs/navbar.html',
     uikit3: 'https://getuikit.com/docs/navbar',
-    pure: 'http://purecss.io/menus/',
-    primer: 'http://primercss.io/nav/',
     semantic: 'http://semantic-ui.com/collections/menu.html',
     spectre: 'https://picturepan2.github.io/spectre/layout.html#navbar'
   }